:hammer: Move print utils to printing.
